working scrim and move traitrow to its own file
